% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spict.R
\name{spict}
\alias{spict}
\alias{SPiCT}
\title{Stochastic production in continuous time (spict)}
\usage{
spict(
  x = 1,
  Data,
  inp_args = list(),
  start = NULL,
  fix_dep = TRUE,
  fix_n = TRUE,
  fix_sigma = FALSE,
  fix_omega = FALSE,
  fix_alpha = TRUE,
  fix_beta = TRUE,
  n_seas = 4L,
  ...
)
}
\arguments{
\item{x}{An index for the objects in \code{Data} when running in \link[DLMtool]{runMSE}.
Otherwise, equals to 1 when running an assessment interactively.}

\item{Data}{An object of class Data.}

\item{inp_args}{A named of list of additional variables for the \code{inp} list to pass to \link[spict]{check.inp}. Core function for configuring spict.}

\item{start}{Optional list of starting values. See details. Overrides \code{inp_args}.}

\item{fix_dep}{Logical, whether to fix the initial depletion (ratio of biomass to carrying capacity in the
first year of the model). If \code{TRUE}, uses the value in \code{start}, otherwise equal to 1
(assumes unfished conditions). Overrides \code{inp_args}.}

\item{fix_n}{Logical, whether to fix the exponent of the production function. If \code{TRUE},
uses the value in \code{start}, otherwise equal to \code{n = 2}, where the biomass at MSY
is half of carrying capacity. Overrides \code{inp_args}.}

\item{fix_sigma}{Logical, whether the standard deviation of the index is fixed. If \code{TRUE},
sigma is fixed to value provided in \code{start} (if provided), otherwise, value based on \code{Data@CV_Ind}. Overrides \code{inp_args}.}

\item{fix_omega}{Logical, whether the standard deviation of the catch is fixed. If \code{TRUE},
omega is fixed to value provided in \code{start} (if provided), otherwise, value based on \code{Data@CV_Cat}. Overrides \code{inp_args}.}

\item{fix_alpha}{Logical, whether the ratio of index and biomass standard deviations is fixed. If \code{TRUE},
alpha is fixed to value provided in \code{start} (if provided), otherwise, equal to 1. Overrides \code{inp_args}.}

\item{fix_beta}{Logical, whether the ratio of catch and F standard deviations is fixed. If \code{TRUE},
tau is fixed to value provided in \code{start} (if provided), otherwise, equal to 1. Overrides \code{inp_args}.}

\item{n_seas}{Integer, the number of sub-annual time steps in the model, i.e. the inverse of the Euler time step. Defaults to 4 (seasonal time steps).}

\item{...}{Additional arguments (not currently used).}
}
\value{
An object of \code{\linkS4class{Assessment}} containing assessment output.
}
\description{
A state-space surplus production model that uses a time-series of catches and a relative abundance index
and coded in TMB (Pedersen and Berg 2017). This is a wrapper assessment function for the \code{SAMtool} package, which
calls the model from the \code{spict} package. There are a multitude of possible configurations for spict. Here, the default is to use
\code{n = 2} (symmetric yield curve), \code{dep = 1} (unfished stock in year 1 of model), and \code{alpha, beta = 1} (equal observation
and process standard deviations). These can be changed in the function arguments below.
}
\details{
To provide starting values, a named list can be provided for \code{MSY}, \code{K}, \code{n}, \code{omega}, and
\code{sigma} via the \code{start} argument. Otherwise, the defaults from the spict package are used.
}
\note{
This is a wrapper function intended for a DLMtool \linkS4class{Data} object. The \code{spict} package can be
downloaded from Github with \code{devtools::install_github("DTUAqua/spict/spict")}.

The full spict model also accommodates time-varying reference points and seasonal data among other
things, but these features are currently not used here. Additional diagnostics and plotting functions are also available in the spict package.
}
\section{Required Data}{

\itemize{
\item \code{spict}: Cat, Ind
}
}

\section{Optional Data}{

\itemize{
\item \code{spict}: CV_Cat, CV_Ind
}
}

\examples{
\donttest{
library(openMSE)
data(swordfish)

res <- spict(Data = swordfish)
plot(res)
summary(res)

## Use additional spict-package functions
spict_output <- res@info
library(spict)
summary(spict_output)
plot(spict_output)
spict_retro <- retro(spict_output)
plotspict.retro(spict_retro)
}
}
\references{
Pedersen, M. W. and Berg, C. W. 2017. A stochastic surplus production model in continuous time. Fish and Fisheries. 18:226-243.
}
\seealso{
\link[SAMtool]{SP_production} \link[SAMtool]{plot.Assessment} \link[SAMtool]{summary.Assessment} \link[SAMtool]{make_MP}
}
\author{
Q. Huynh
}
