% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{MSEextra-OM}
\alias{MSEextra-OM}
\alias{MSEextra}
\alias{MSEextra-package}
\alias{Almaco_Jack_GOM_NOAA}
\alias{Arrowtooth_Flounder_BC_DFO}
\alias{Atlantic_Herring_4TFallSpawn_DFO}
\alias{Atlantic_Herring_4TSpringSpawn_DFO}
\alias{Banded_Morwong_Tasmania}
\alias{Barred_Sand_Bass_CA_CDFW}
\alias{California_Halibut_CA_CDFW}
\alias{Canary_Rockfish_BC_DFO}
\alias{Cod_4T4Vn_DFO}
\alias{Cod_4X5Y_DFO}
\alias{Cod_5ZJM_DFO}
\alias{Gatuzo_Argentina_INIDEP}
\alias{Haddock_4X5Y_DFO}
\alias{Haddock_Faroe_Islands}
\alias{Hogfish_PR_NOAA}
\alias{Jonah_Crab_LFA34_DFO}
\alias{Kelp_Bass_CA_CDFW}
\alias{Lane_Snapper_GOM_NOAA}
\alias{Lesser_Amberjack_GOM_NOAA}
\alias{Pacific_Hake_USCan}
\alias{Pacific_Ocean_Perch_QC_BC_DFO}
\alias{Queen_Triggerfish_STT_NOAA}
\alias{Red_Drum_GOM_NOAA}
\alias{Red_Sea_Urchin_CA_CDFW}
\alias{Redtail_Surf_Perch_CA_CDFW}
\alias{Rock_Crab_CA_CDFW}
\alias{Shortspine_Thornyhead_BC_DFO}
\alias{Skipjack_IO_IOTC}
\alias{Snowy_Grouper_GOM_NOAA}
\alias{Speckled_Hind_GOM_NOAA}
\alias{Spiny_Lobster_CA_CDFW}
\alias{Spiny_Lobster_STX_NOAA}
\alias{Stones_Sheep_BC}
\alias{Stoplight_Parrotfish_STX_NOAA}
\alias{Warty_Sea_Cucumber_CA_CDFW}
\alias{Yelloweye_Rockfish_BC_DFO}
\alias{Yellowmouth_Grouper_GOM_NOAA}
\alias{Yellowtail_Snapper_PR_NOAA}
\title{Operating models and MSE output for openMSE}
\format{
An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.
}
\usage{
Almaco_Jack_GOM_NOAA

Arrowtooth_Flounder_BC_DFO

Atlantic_Herring_4TFallSpawn_DFO

Atlantic_Herring_4TSpringSpawn_DFO

Banded_Morwong_Tasmania

Barred_Sand_Bass_CA_CDFW

California_Halibut_CA_CDFW

Canary_Rockfish_BC_DFO

Cod_4T4Vn_DFO

Cod_4X5Y_DFO

Cod_5ZJM_DFO

Gatuzo_Argentina_INIDEP

Haddock_4X5Y_DFO

Haddock_Faroe_Islands

Hogfish_PR_NOAA

Jonah_Crab_LFA34_DFO

Kelp_Bass_CA_CDFW

Lane_Snapper_GOM_NOAA

Lesser_Amberjack_GOM_NOAA

Pacific_Hake_USCan

Pacific_Ocean_Perch_QC_BC_DFO

Queen_Triggerfish_STT_NOAA

Red_Drum_GOM_NOAA

Red_Sea_Urchin_CA_CDFW

Redtail_Surf_Perch_CA_CDFW

Rock_Crab_CA_CDFW

Shortspine_Thornyhead_BC_DFO

Skipjack_IO_IOTC

Snowy_Grouper_GOM_NOAA

Speckled_Hind_GOM_NOAA

Spiny_Lobster_CA_CDFW

Spiny_Lobster_STX_NOAA

Stones_Sheep_BC

Stoplight_Parrotfish_STX_NOAA

Warty_Sea_Cucumber_CA_CDFW

Yelloweye_Rockfish_BC_DFO

Yellowmouth_Grouper_GOM_NOAA

Yellowtail_Snapper_PR_NOAA
}
\description{
A suite of operating models (class OM)
}
\examples{
library(openMSE)
library(MSEextra)
myMSE <- runMSE(Shortspine_Thornyhead_BC_DFO)
plot(Shortspine_Thornyhead_BC_DFO)
}
\references{
\url{}
}
\keyword{datasets}
